moved editTargetVerse into Container so it can be used in verse editor dialog.  fix unit tests
